Finish Kwalify schema definition for the masterlist format.
As mentioned in the last commit, Kwalify doesn't support a value to by
either one thing (e.g. a string) or another thing (e.g., a mapping),
which means that for those items that can be either a string or a
sequence of mappings, I have now settled for just using `type: any`.
Even though this does decrease the effectiveness at catching issues with
the masterlists, the schema as it is is still far more in-depth than
simply checking for YAML validity.

Speaking of, this file is valid YAML, but `pkwalify` doesn't seem to
like it. I haven't tested with the Ruby `kwalify` yet.

Part of issue #18.