Add page redirecting to the latest LOOT thread.
This will allow us to only update the LOOT thread link in one place
(well, three as per this commit, but they're all in the same file :)),
which should lead to a much reduced chance of masterlists becoming
obsolete across games and versions.

The only downside (that I can think of) is that browsers (including
LOOT's CEF) will no longer immediately recognise that the new thread
has not been visited. I feel that this is insignificant, as users could
already not see, based on the link, whether there had been new replies
or other important info posted to the thread. If people want to keep
that up to date with the thread, Bethesda's forum has a way to subscribe
to threads.

See also:
- https://help.github.com/articles/redirects-on-github-pages/
- https://github.com/jekyll/jekyll-redirect-from