Link to "a pastebin" instead of Pastebin(.com).
Replaced the link to Pastebin.com with a link to Pastie.org.
Unfortunately not an HTTPS one, but the HTTPS ones either had a very
short default Time-To-Live on their pastes or were not very non-techie
friendly/oriented. :/

Pastie.org is very simple: it has a (large) field for copying in the log
file (or other data) and then you can submit. You can also select
language and make the paste be "private" (basically, it won't get
Google indexed) - but there's no obligation to worry about either
of those options.

https://wiki.archlinux.org/index.php/List_of_applications#Pastebin_clients :
> Note: pastebin.com is blocked for some people and has a history of
> annoying issues (javascript, adverts, poor formatting, etc).