Use a table to simplify instructions
The current instructions are quite a mess, so tidy them up with an HTML
table. We can also utilize the table to do away with the link/path
replacement & just list each link & path explicitly in a much neater
format. The note can be removed too since the links aren't directly
exposed to the user, avoiding most of the potential confusion it was
attempting to address.